Break the configure model into several distinct files.
